+++ src/ZoomButton.js
@@ -1,20 +1,20 @@
-import videojs from 'video.js';
+import videojs from "video.js";
 
-const Button = videojs.getComponent('Button');
+const Button = videojs.getComponent("Button");
 
 export class ZoomButton extends Button {
 
 	constructor(player, options) {
 		super(player, options);
 		this.isOpen = false;
-		player.on('useractive', () => {
+		player.on("useractive", () => {
 			if (!this.isOpen) return;
-			const modal = this.player().getChild('ZoomModal');
+			const modal = this.player().getChild("ZoomModal");
 			modal.open();
 		});
-		player.on('userinactive', () => {
+		player.on("userinactive", () => {
 			if (!this.isOpen) return;
-			const modal = this.player().getChild('ZoomModal');
+			const modal = this.player().getChild("ZoomModal");
 			modal.close();
 		});
 	}
@@ -24,8 +24,8 @@ export class ZoomButton extends Button {
 	}
 
 	handleClick() {
-		const modal = this.player().getChild('ZoomModal');
-		videojs.log('[~Zoom Plugin] button handleClick');
+		const modal = this.player().getChild("ZoomModal");
+		videojs.log("[~Zoom Plugin] button handleClick");
 		this.isOpen = !this.isOpen;
 		modal.toggle();
 	}