+++ src/ZoomFunction.js
@@ -8,10 +8,10 @@ export class ZoomFunction {
 		this.player = player.el();
 		this.plugin = options.plugin;
 		this.observer = Observer.getInstance();
-		player.on('playing', () => {
+		player.on("playing", () => {
 			this._updateSalt();
 		});
-		this.observer.subscribe('change', state => {
+		this.observer.subscribe("change", state => {
 			this.state = { ...state, saltMoveX: 70, saltMoveY: 70 };
 			this._updateSalt();
 		});